UBUNTU: include: configs: Use SiFive Unmatched board ennvironment
Use the SiFive Unmatched environment that we already use for the Ubuntu
image we deliver on this board, it is then easier to maintain.

In addition, for now, I don't see the point of having a uEnv.txt file in
the /boot partition, it is easier to have everything in the config file.

Signed-off-by: Alexandre Ghiti <alexandre.ghiti@canonical.com>